add browser support
