+++ forward_engineering/api.js
@@ -1,80 +1,17 @@
-const _ = require('lodash');
-const reApi = require('../reverse_engineering/api');
-const { createLogger } = require('../reverse_engineering/helpers/loggerHelper');
-const applyToInstanceHelper = require('./applyToInstanceHelper');
-const {
-	buildEntityLevelAlterScript,
-	buildContainerLevelAlterScript,
-	doesContainerLevelAlterScriptContainDropStatements,
-	doesEntityLevelAlterScriptContainDropStatements,
-} = require('./alterScript/alterScriptBuilder');
+const { generateScript } = require('./api/generateScript');
+const { generateViewScript } = require('./api/generateViewScript');
+const { generateContainerScript } = require('./api/generateContainerScript');
+const { getDatabases } = require('./api/getDatabases');
+const { applyToInstance } = require('./api/applyToInstance');
+const { testConnection } = require('./api/testConnection');
+const { isDropInStatements } = require('./api/isDropInStatements');
 
 module.exports = {
-	generateScript(data, logger, callback, app) {
-		try {
-			const script = buildEntityLevelAlterScript(data, app);
-			callback(null, script);
-		} catch (error) {
-			logger.log('error', { message: error.message, stack: error.stack }, 'PostgreSQL Forward-Engineering Error');
-
-			callback({ message: error.message, stack: error.stack });
-		}
-	},
-
-	generateViewScript(data, logger, callback, app) {
-		callback(new Error('Forward-Engineering of delta model on view level is not supported'));
-	},
-
-	generateContainerScript(data, logger, callback, app) {
-		try {
-			const script = buildContainerLevelAlterScript(data, app);
-			callback(null, script);
-		} catch (error) {
-			logger.log('error', { message: error.message, stack: error.stack }, 'PostgreSQL Forward-Engineering Error');
-
-			callback({ message: error.message, stack: error.stack });
-		}
-	},
-
-	getDatabases(connectionInfo, logger, callback, app) {
-		logger.progress({ message: 'Find all databases' });
-
-		reApi.getDatabases(connectionInfo, logger, callback, app);
-	},
-
-	applyToInstance(connectionInfo, logger, callback, app) {
-		logger.clear();
-		logger.log(
-			'info',
-			_.omit(connectionInfo, 'script', 'containerData'),
-			'connectionInfo',
-			connectionInfo.hiddenKeys,
-		);
-
-		const postgresLogger = createLogger({
-			title: 'Apply to instance',
-			hiddenKeys: connectionInfo.hiddenKeys,
-			logger,
-		});
-
-		applyToInstanceHelper.applyToInstance(connectionInfo, postgresLogger, app).then(callback, callback);
-	},
-
-	testConnection(connectionInfo, logger, callback, app) {
-		reApi.testConnection(connectionInfo, logger, callback, app).then(callback, callback);
-	},
-
-	isDropInStatements(data, logger, callback, app) {
-		try {
-			if (data.level === 'container') {
-				const containsDropStatements = doesContainerLevelAlterScriptContainDropStatements(data, app);
-				callback(null, containsDropStatements);
-			} else {
-				const containsDropStatements = doesEntityLevelAlterScriptContainDropStatements(data, app);
-				callback(null, containsDropStatements);
-			}
-		} catch (e) {
-			callback({ message: e.message, stack: e.stack });
-		}
-	},
+	generateScript,
+	generateViewScript,
+	generateContainerScript,
+	getDatabases,
+	applyToInstance,
+	testConnection,
+	isDropInStatements,
 };
