+++ forward_engineering/api/getDatabases.js
@@ -1,0 +1,11 @@
+const reApi = require('../../reverse_engineering/api');
+
+function getDatabases(connectionInfo, logger, callback, app) {
+	logger.progress({ message: 'Find all databases' });
+
+	reApi.getDatabases(connectionInfo, logger, callback, app);
+}
+
+module.exports = {
+	getDatabases,
+};
